# apps/api/openapi.gen.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Skald AI API",
    "description": "AI services API for speech-to-text transcription, LLM chat completions, and subscription management",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/llm/chat/completions": {
      "post": {
        "tags": ["llm"],
        "operationId": "llm_chat_completions",
        "parameters": [
          {
            "name": "x-char-task",
            "in": "header",
            "description": "Task type for model selection",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/SkaldTask"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat completion response (streaming or non-streaming)"
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "502": {
            "description": "Upstream provider failed"
          },
          "504": {
            "description": "Request timeout"
          }
        }
      }
    },
...
    "/stt/listen": {
      "get": {
        "tags": ["stt"],
        "operationId": "stt_listen_stream",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "description": "STT provider. Use 'skald' for automatic routing (default), or specify:\ndeepgram, soniox, assemblyai, gladia, elevenlabs, fireworks, openai, dashscope, mistral",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model",
            "in": "query",
            "description": "Model to use for transcription (provider-specific, e.g. 'nova-3')",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "query",
            "description": "BCP-47 language hint (e.g. 'en', 'ko', 'ja'). Multiple values or comma-separated supported",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keywords",
            "in": "query",
            "description": "Keyword boosting. Comma-separated or repeated query params",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sample_rate",
            "in": "query",
            "description": "Audio sample rate in Hz (default: 16000)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "channels",
            "in": "query",
            "description": "Number of audio channels (default: 1)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "encoding",
            "in": "query",
            "description": "Audio encoding: linear16, flac, mulaw, opus, ogg-opus, etc.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "WebSocket upgrade. Server sends StreamResponse JSON messages."
          },
          "400": {
            "description": "Bad request (invalid provider, missing params, etc.)"
          },
          "502": {
            "description": "Upstream provider connection failed"
          }
        }
      },
      "post": {
        "tags": ["stt"],
        "operationId": "stt_listen_batch",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "description": "STT provider. Use 'skald' for automatic routing (default), or specify:\ndeepgram, soniox, assemblyai, gladia, elevenlabs, fireworks, openai, dashscope, mistral",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model",
            "in": "query",
            "description": "Model to use for transcription (provider-specific, e.g. 'nova-3')",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "query",
            "description": "BCP-47 language hint (e.g. 'en', 'ko', 'ja'). Multiple values or comma-separated supported",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keywords",
            "in": "query",
            "description": "Keyword boosting. Comma-separated or repeated query params",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "callback",
            "in": "query",
            "description": "When set, enables async callback mode. Body should be JSON with a `url` field instead of raw audio",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Raw audio bytes (sync mode) or JSON `{ \"url\": \"<file_id>\" }` (callback mode)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListenCallbackRequest"
              }
            },
            "application/octet-stream": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "binary",
                  "minimum": 0
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Transcription result (sync) or ListenCallbackResponse (callback)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchListenSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (empty body, invalid provider, etc.)"
          },
          "401": {
            "description": "Unauthorized"
          },
          "502": {
            "description": "All upstream providers failed"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchAlternatives": {
        "type": "object",
        "required": ["transcript", "confidence"],
        "properties": {
          "confidence": {
            "type": "number",
            "format": "double"
          },
          "transcript": {
            "type": "string"
          },
          "words": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BatchWord"
            }
          }
        }
      },
      "BatchChannel": {
        "type": "object",
        "required": ["alternatives"],
        "properties": {
          "alternatives": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BatchAlternatives"
            }
          }
        }
      },
      "BatchListenSuccessResponse": {
        "oneOf": [
          {
            "type": "object",
            "required": ["Sync"],
            "properties": {
              "Sync": {
                "$ref": "#/components/schemas/BatchResponse"
              }
            }
          },
          {
            "type": "object",
            "required": ["Callback"],
            "properties": {
              "Callback": {
                "$ref": "#/components/schemas/ListenCallbackResponse"
              }
            }
          }
        ]
      },
      "BatchResponse": {
        "type": "object",
        "required": ["metadata", "results"],
        "properties": {
          "metadata": {
            "type": "object"
          },
          "results": {
            "$ref": "#/components/schemas/BatchResults"
          }
        }
      },
      "BatchResults": {
        "type": "object",
        "required": ["channels"],
        "properties": {
          "channels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BatchChannel"
            }
          }
        }
      },
...
      "BatchWord": {
        "type": "object",
        "required": ["word", "start", "end", "confidence"],
        "properties": {
          "channel": {
            "type": "integer",
            "format": "int32"
          },
          "confidence": {
            "type": "number",
            "format": "double"
          },
          "end": {
            "type": "number",
            "format": "double"
          },
          "punctuated_word": {
            "type": ["string", "null"]
          },
          "speaker": {
            "type": ["integer", "null"],
            "minimum": 0
          },
          "start": {
            "type": "number",
            "format": "double"
          },
          "word": {
            "type": "string"
          }
        }
      },
...
      "SkaldTask": {
        "type": "string",
        "enum": ["chat", "enhance", "title"]
      },
...
      "ListenCallbackRequest": {
        "type": "object",
        "required": ["url"],
        "properties": {
          "url": {
            "type": "string"
          }
        }
      },
      "ListenCallbackResponse": {
        "type": "object",
        "required": ["request_id"],
        "properties": {
          "request_id": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token"
      }
    }
  },
...
}